jwt인증 인터셉터, jwt토큰 아규먼트리졸버, 인가기능, 예외처리 추가
* jwt토큰 인증을 위한 인터셉터 추가
* jwt토큰을 UserDto로 바인딩하는 아규먼트리졸버 추가
* 상품수정, 관심상품등록, 관심상품삭제시 인가기능 추가
* ControllerAdvice를 통한 예외처리 추가

+++ product/src/main/java/com/hanium/product/service/impl/ProductServiceImpl.java
@@ -6,6 +6,8 @@ import com.hanium.product.dao.IProductImageDao;
 import com.hanium.product.dto.ProductArticleDto;
 import com.hanium.product.dto.ProductArticleRequestDto;
 import com.hanium.product.dto.ProductImageDto;
+import com.hanium.product.exception.AuthorizationException;
+import com.hanium.product.exception.NotSavedFileException;
 import com.hanium.product.service.ProductService;
 import lombok.AllArgsConstructor;
 import org.springframework.beans.factory.annotation.Value;
@@ -41,11 +43,12 @@ public class ProductServiceImpl implements ProductService {
 
     @Transactional
     @Override
-    public Integer createProductArticle(ProductArticleDto.Info productArticle, List<MultipartFile> multipartFiles) throws Exception {
+    public Integer createProductArticle(ProductArticleDto.Info productArticle, List<MultipartFile> multipartFiles) {
         productArticleDao.createBy(productArticle);
+
         List<ProductImageDto> productImages = fileUtils.parseImageInfo(productArticle.getId(), multipartFiles);
         if(productImages == null) {
-            throw new RuntimeException("NotSavedImageFiles");
+            throw new NotSavedFileException("상품이미지를 저장 할 수 없습니다.");
         }
         productImageDao.createList(productImages);
         return productArticle.getId();
@@ -53,16 +56,25 @@ public class ProductServiceImpl implements ProductService {
 
     @Transactional
     @Override
-    public void updateProductArticle(ProductArticleDto.ChangeInfo changeInfo, Integer articleId) throws Exception {
+    public void updateProductArticle(ProductArticleDto.ChangeInfo changeInfo, Integer articleId, Integer userId) {
+        ProductArticleDto.Info article = productArticleDao.findBy(articleId);
+        if(!article.getUser().getId().equals(userId)){
+            throw new AuthorizationException("게시글 작성자만 글을 수정할 수 있습니다.");
+        }
+
         productArticleDao.updateBy(changeInfo, articleId);
         productImageDao.deleteBy(articleId);
+
         List<ProductImageDto> productImages = fileUtils.parseImageInfo(articleId, changeInfo.getFiles());
+        if(productImages == null){
+            throw new NotSavedFileException("상품이미지를 저장 할 수 없습니다.");
+        }
         productImageDao.createList(productImages);
     }
 
     @Override
-    public void deleteProductArticle(Integer id) {
-        productArticleDao.deleteBy(id);
+    public void deleteProductArticle(Integer articleId, Integer userId) {
+        productArticleDao.deleteBy(articleId);
     }
 
 }
